+++ lib/app_security_kit.dart
@@ -1,2 +1,3 @@
 export 'package:app_security_kit/retrieve.dart';
 export 'package:app_security_kit/rsa_pem.dart';
+export 'package:app_security_kit/password_encrypt_helper.dart';