+++ CHANGELOG.md
@@ -1,2 +1,5 @@
+## 0.0.2
++ Added Password encryption
+    + Encrypt / Decrypt string with password
 ## 0.0.1
 * Initialize Package